refactor(e2e): split toolkit/adapters.py into adapters + builders + ci_lanes
Pure relocation, no behavior change. adapters.py keeps the registry core
(Adapter, AdapterSpec, @adapter/_REGISTRY, spec_for, tool helpers, discovery
guard, adapter_lane, specs, build_adapter); builders.py holds the 12 _build_*
functions (registered via @adapter import side-effect, triggered by a deferred
bottom import in adapters.py); ci_lanes.py holds the CI-lane partition +
workflow-drift guards (CILane, ci_lanes, assert_every_adapter_has_a_ci_home,
workflow_lane_* + asserts). Import graph is acyclic. Import sites and the e2e.yml
inline block updated to source lane/workflow names from ci_lanes.
