+++ tests/e2e/baseline/toolkit/builders.py
@@ -1,0 +1,327 @@
+"""The matrix: one self-registering builder per LLM-agent adapter (pytest-free).
+
+Each builder lazy-imports its framework and maps the generic ``prompt`` to the
+constructor argument that framework uses (prompt / custom_section / system_prompt /
+the agent's own instructions). ``supports`` lists the platform capabilities the
+adapter advertises for capability-scoped matrices.
+
+This module has no public API: importing it runs the ``@adapter`` decorators, which
+populate the registry in ``adapters``. ``adapters`` imports it once (at the bottom of
+that module) so the registry is populated before ``specs()`` / ``build_adapter`` query
+it. Heavy/optional framework imports live **inside** each builder so importing this
+module never pulls in an absent dependency.
+
+To add a framework: add an ``Adapter`` enum member in ``adapters`` and a decorated
+builder here (see ``adapters`` module docstring for the full recipe).
+"""
+
+from __future__ import annotations
+
+from typing import Any
+
+from band.core.simple_adapter import SimpleAdapter
+from band.core.types import AdapterFeatures, Capability
+
+from tests.e2e.baseline.settings import BaselineSettings
+from tests.e2e.baseline.toolkit.adapters import (
+    Adapter,
+    _custom_tool_defs,
+    _reject_tools,
+    adapter,
+)
+from tests.e2e.baseline.toolkit.requirements import Dep
+from tests.e2e.baseline.toolkit.tools import ToolSpec
+
+_LLM_TOOL_LOOP = (Capability.MEMORY, Capability.CONTACTS)
+
+
+@adapter(Adapter.ANTHROPIC, requires=[Dep.ANTHROPIC], supports=_LLM_TOOL_LOOP)
+def _build_anthropic(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    from band.adapters.anthropic import AnthropicAdapter
+
+    return AnthropicAdapter(
+        model=s.llm_models.anthropic_model,
+        provider_key=s.llm_credentials.anthropic_api_key or None,
+        prompt=prompt,
+        additional_tools=_custom_tool_defs(tools),
+        features=features,
+    )
+
+
+@adapter(Adapter.CLAUDE_SDK, requires=[Dep.ANTHROPIC], supports=_LLM_TOOL_LOOP)
+def _build_claude_sdk(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    from band.adapters.claude_sdk import ClaudeSDKAdapter
+
+    return ClaudeSDKAdapter(
+        model=s.llm_models.anthropic_model,
+        custom_section=prompt,
+        additional_tools=_custom_tool_defs(tools),
+        features=features,
+    )
+
+
+@adapter(Adapter.LANGGRAPH, requires=[Dep.OPENAI], supports=_LLM_TOOL_LOOP)
+def _build_langgraph(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    from langchain_openai import ChatOpenAI
+    from langgraph.checkpoint.memory import MemorySaver
+
+    from band.adapters.langgraph import LangGraphAdapter
+
+    return LangGraphAdapter(
+        llm=ChatOpenAI(
+            model=s.llm_models.openai_model,
+            api_key=s.llm_credentials.openai_api_key or None,
+        ),
+        # Deliberately an in-memory checkpointer: it is rebuilt fresh on every
+        # cell.run_as, so no LangGraph state survives a reboot in-process. That is
+        # what keeps the rehydration scenarios honest for this cell — recall after a
+        # reboot can only come from platform /context, not the checkpointer. Swapping
+        # in a persistent checkpointer keyed by room_id would silently move langgraph
+        # into the codex/opencode "backend session resume" class and invalidate that.
+        checkpointer=MemorySaver(),
+        custom_section=prompt or "",
+        additional_tools=_custom_tool_defs(tools),
+        features=features,
+    )
+
+
+@adapter(Adapter.PYDANTIC_AI, requires=[Dep.OPENAI], supports=_LLM_TOOL_LOOP)
+def _build_pydantic_ai(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    from pydantic_ai import RunContext
+
+    from band.adapters.pydantic_ai import PydanticAIAdapter
+
+    # pydantic-ai takes native callables with a RunContext-first signature.
+    native = (
+        [t.as_callable(ctx_annotation=RunContext) for t in tools] if tools else None
+    )
+    return PydanticAIAdapter(
+        model=f"openai:{s.llm_models.openai_model}",
+        custom_section=prompt,
+        additional_tools=native,
+        features=features,
+    )
+
+
+@adapter(Adapter.GEMINI, requires=[Dep.GOOGLE], supports=_LLM_TOOL_LOOP)
+def _build_gemini(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    from band.adapters.gemini import GeminiAdapter
+
+    return GeminiAdapter(
+        model=s.llm_models.gemini_model,
+        provider_key=s.llm_credentials.google_api_key or None,
+        prompt=prompt,
+        additional_tools=_custom_tool_defs(tools),
+        features=features,
+    )
+
+
+@adapter(Adapter.GOOGLE_ADK, requires=[Dep.GOOGLE], supports=_LLM_TOOL_LOOP)
+def _build_google_adk(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    from band.adapters.google_adk import GoogleADKAdapter
+
+    # google-adk reads the provider key / Vertex config from the environment.
+    return GoogleADKAdapter(
+        model=s.llm_models.gemini_model,
+        custom_section=prompt,
+        additional_tools=_custom_tool_defs(tools),
+        features=features,
+    )
+
+
+@adapter(Adapter.CREWAI, requires=[Dep.OPENAI, Dep.CREWAI], supports=_LLM_TOOL_LOOP)
+def _build_crewai(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    from band.adapters.crewai import CrewAIAdapter
+
+    return CrewAIAdapter(
+        model=s.llm_models.openai_model,
+        role="Test Assistant",
+        goal="Help users with simple tasks for testing.",
+        backstory="A test agent for E2E validation.",
+        custom_section=prompt,
+        additional_tools=_custom_tool_defs(tools),
+        features=features,
+    )
+
+
+@adapter(Adapter.AGNO, requires=[Dep.ANTHROPIC], supports=_LLM_TOOL_LOOP)
+def _build_agno(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    # Agno bridges a user-built agent, so steering goes into its instructions.
+    # Use the Anthropic model: small models refuse the suite's crafted prompts as
+    # injection, so the matrix relies on E2E_ANTHROPIC_MODEL being a capable model.
+    from agno.agent import Agent as AgnoAgent
+    from agno.models.anthropic import Claude
+
+    from band.adapters.agno import AgnoAdapter
+
+    # agno tools are plain callables on the agent; the band adapter captures them
+    # and re-offers them alongside the platform tools each run.
+    native = [t.as_callable() for t in tools] if tools else None
+    return AgnoAdapter(
+        AgnoAgent(
+            model=Claude(id=s.llm_models.anthropic_model),
+            instructions=prompt,
+            tools=native,
+        ),
+        features=features,
+    )
+
+
+@adapter(Adapter.CREWAI_FLOW, requires=[Dep.CREWAI], runs_tool_loop=False)
+def _build_crewai_flow(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    # CrewAI Flow returns a terminal result rather than running the Band tool loop,
+    # so it takes a flow_factory (not a model/prompt) and advertises no platform
+    # capabilities. The minimal flow echoes back so the reply path is observable.
+    from band.adapters.crewai_flow import CrewAIFlowAdapter
+
+    class _E2EFlow:
+        async def kickoff_async(self, inputs: dict[str, Any]) -> dict[str, Any]:
+            message = inputs.get("message", {})
+            content = message.get("content", "") if isinstance(message, dict) else ""
+            return {"decision": "direct_response", "content": content, "mentions": []}
+
+    return CrewAIFlowAdapter(
+        flow_factory=_E2EFlow,
+        additional_tools=_custom_tool_defs(tools),
+        features=features,
+    )
+
+
+@adapter(Adapter.CODEX, requires=[Dep.CODEX_CLI, Dep.CODEX_CWD], runs_tool_loop=False)
+def _build_codex(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    from band.adapters.codex import CodexAdapter, CodexAdapterConfig
+
+    # Only override what's explicitly configured. CODEX_MODEL is left unset by
+    # default -- NOT defaulted to the OpenAI chat model: Codex uses its own model
+    # catalogue (gpt-4o-mini isn't in it), so leaving config.model=None lets the
+    # adapter discover/select a valid Codex model. CODEX_COMMAND likewise: an absent
+    # value spawns the stock `codex` binary. Splits mirror the gates in requirements.py.
+    config_kwargs: dict[str, Any] = {
+        "cwd": s.backends.codex_cwd,
+        "custom_section": prompt or "",
+    }
+    if s.backends.codex_model.strip():
+        config_kwargs["model"] = s.backends.codex_model
+    if s.backends.codex_command.strip():
+        config_kwargs["codex_command"] = tuple(s.backends.codex_command.split())
+
+    return CodexAdapter(
+        config=CodexAdapterConfig(**config_kwargs),
+        additional_tools=_custom_tool_defs(tools),
+        features=features,
+    )
+
+
+@adapter(Adapter.OPENCODE, requires=[Dep.OPENCODE_SERVER], runs_tool_loop=False)
+def _build_opencode(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    from band.adapters.opencode import OpencodeAdapter, OpencodeAdapterConfig
+
+    return OpencodeAdapter(
+        config=OpencodeAdapterConfig(
+            base_url=s.backends.opencode_base_url,
+            provider_id=s.backends.opencode_provider_id,
+            model_id=s.backends.opencode_model_id,
+            custom_section=prompt or "",
+        ),
+        additional_tools=_custom_tool_defs(tools),
+        features=features,
+    )
+
+
+# Letta is registered so the `letta` CI lane is still defined (ci_lanes), but
+# e2e_pending=True drops it from specs()/adapter_params(), so the matrix runs no
+# cells for it. (e2e_pending does not gate @with_adapters — there simply are no
+# @with_adapters(Adapter.LETTA) tests.) Letta executes platform tools only by calling
+# a band-mcp server, and standing one up reachable from the Letta server (its SSRF
+# guard rejects loopback) isn't wired yet, so the live smokes are deferred.
+#
+# TODO: cover Letta live once a reachable band-mcp + the Letta smokes land — flip
+# e2e_pending to False below (or drop the kwarg) to return it to the matrix, and
+# add the smokes; nothing else here changes.
+@adapter(Adapter.LETTA, requires=[Dep.LETTA], e2e_pending=True, runs_tool_loop=False)
+def _build_letta(
+    s: BaselineSettings,
+    *,
+    prompt: str | None,
+    features: AdapterFeatures | None,
+    tools: list[ToolSpec] | None = None,
+) -> SimpleAdapter[Any]:
+    from band.adapters.letta import LettaAdapter, LettaAdapterConfig
+
+    _reject_tools(Adapter.LETTA, tools)
+
+    return LettaAdapter(
+        config=LettaAdapterConfig(
+            base_url=s.backends.letta_base_url,
+            provider_key=s.backends.letta_api_key or None,
+            model=s.backends.letta_model,
+            custom_section=prompt or "",
+        ),
+        features=features,
+    )
